refactor: 调整原先 tooltip Action 的位置
